pma/requirements.txt
streamlit
mysql-connector-python
bcrypt
openai
python-dotenv
googlemaps
